% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corteidh_baixar_sentencas.R
\name{corteidh_baixar_sentencas}
\alias{corteidh_baixar_sentencas}
\title{Baixa decisões da Corte IDH}
\usage{
corteidh_baixar_sentencas(urls, diretorio)
}
\arguments{
\item{urls}{Vetor de urls}

\item{diretorio}{Diretório onde armazenar os pdfs}
}
\value{
pdf
}
\description{
Baixa decisões da Corte IDH
}
