% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corteidh_listar_sentencas.R
\name{corteidh_listar_sentencas}
\alias{corteidh_listar_sentencas}
\title{Lista todas as decisóes da Corte de IDH}
\usage{
corteidh_listar_sentencas()
}
\value{
Tibble com título e url para acesso ao pdf
}
\description{
Lista todas as decisóes da Corte de IDH
}
